requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.25
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug>=3.0.1
Flask-Migrate>=4.0.0
psycopg2-binary==2.9.9
